Function to create MBR-design.

Function to create multi-level binary replacement (MBR) design (Martens et al., 2010). The MBR approach was developed for constructing experimental designs for computer experiments. MBR makes it possible to set up fractional designs for multi-factor problems with potentially many levels for each factor. In this package it is mainly called by the mbrdsim function.

mbrd(
  l2levels = c(2, 2),
  fraction = 0,
  gen = NULL,
  fnames1 = NULL,
  fnames2 = NULL
)

Arguments

l2levels
A vector indicating the number of log2-levels for each factor. E.g. c(2,3) means 2 factors, the first with \(2^2=4\) levels, the second with \(2^3=8\) levels

fraction
Design fraction at bit-level. Full design: fraction=0, half-fraction: fraction=1, and so on...

gen
list of generators at bit-factor level. Same as generators in function FrF2.

fnames1
Factor names of original multi-level factors (optional).

fnames2
Factor names at bit-level (optional).

Value

BitDesign

The design at bit-factor level (inherits from FrF2). Function design.info() can be used to get extra design info of the bit-design, and plot for plotting of the bit-level design.

Design

The design at original factor levels, non-randomized.

Details

The MBR design approach was developed for designing fractional designs in multi-level multi-factor experiments, typically computer experiments. The basic idea can be summarized in the following steps: 1) Choose the number of levels \(L\) for each multi-level factor as a multiple of 2, that is \(L \in \{2, 4, 8,...\}\). 2) Replace any given multi-level factor by a set of \(ln(L)\) two-level "bit factors". The complete bit-factor design can then by expressed as a \(2^K\) design where \(K\) is the total number of bit-factors across all original multi-level factors. 3) Choose a fraction level \(P\) defining av fractional design \(2^{(K-P)}\) (see e.g. Montgomery, 2008) as for regular two-levels factorial designs. 4) Express the reduced design in terms of the original multi-level factors.

References

Martens, H., Måge, I., Tøndel, K., Isaeva, J., Høy, M. and Sæbø¸, S., 2010, Multi-level binary replacement (MBR) design for computer experiments in high-dimensional nonlinear systems, J, Chemom, 24, 748--756.

Montgomery, D., Design and analysis of experiments, John Wiley & Sons, 2008.

Examples

 #Two variables with 8 levels each (2^3=8), a half-fraction design.
 res <- mbrd(c(3,3),fraction=1, gen=list(c(1,4)))
 #plot(res$Design, pch=20, cex=2, col=2)
 #Three variabler with 8 levels each, a 1/16-fraction.
 res <- mbrd(c(3,3,3),fraction=4)
 #library(rgl)
 #plot3d(res$Design,type="s",col=2)