Simulation of Multivariate Linear Model Data

Simulation of Multivariate Linear Model Data

msim(
  p = 15,
  q = c(5, 4, 3),
  m = 5,
  relpos = list(c(1, 2), c(3, 4, 6), c(5, 7)),
  gamma = 0.6,
  R2 = c(0.8, 0.7, 0.8),
  eta = 0,
  muX = NULL,
  muY = NULL,
  ypos = list(c(1), c(3, 4), c(2, 5))
)

Arguments

p
Number of variables

q
Vector containing the number of relevant predictor variables for each relevant response components

m
Number of response variables

relpos
A list of position of relevant component for predictor variables. The list contains vectors of position index, one vector or each relevant response components

gamma
A declining (decaying) factor of eigen value of predictors (X). Higher the value of gamma, the decrease of eigenvalues will be steeper

R2
Vector of coefficient of determination (proportion of variation explained by predictor variable) for each relevant response components

eta
A declining (decaying) factor of eigenvalues of response (Y). Higher the value of eta, more will be the declining of eigenvalues of Y. eta = 0 refers that all eigenvalues of responses (Y) are 1.

muX
Vector of average (mean) for each predictor variable

muY
Vector of average (mean) for each response variable

ypos
List of position of relevant response components that are combined to generate response variable during orthogonal rotation

Value

A simrel object with all the input arguments along with following additional items

X

Simulated predictors

Y

Simulated responses

W

Simulated predictor components

Z

Simulated response components

beta

True regression coefficients

beta0

True regression intercept

relpred

Position of relevant predictors

testX

Test Predictors

testY

Test Response

testW

Test predictor components

testZ

Test response components

minerror

Minimum model error

Xrotation

Rotation matrix of predictor (R)

Yrotation

Rotation matrix of response (Q)

type

Type of simrel object univariate or multivariate

lambda

Eigenvalues of predictors

SigmaWZ

Variance-Covariance matrix of components of response and predictors

SigmaWX

Covariance matrix of response components and predictors

SigmaYZ

Covariance matrix of response and predictor components

Sigma

Variance-Covariance matrix of response and predictors

RsqW

Coefficient of determination corresponding to response components

RsqY

Coefficient of determination corresponding to response variables

References

Sæbø, S., Almøy, T., & Helland, I. S. (2015). simrel—A versatile tool for linear model data simulation based on the concept of a relevant subspace and relevant predictors. Chemometrics and Intelligent Laboratory Systems, 146, 128-135.

Almøy, T. (1996). A simulation study on comparison of prediction methods when only a few components are relevant. Computational statistics & data analysis, 21(1), 87-107.