Simulation of Multivariate Linear Model Data

Simulation of Multivariate Linear Model Data

simrel(n, p, q, relpos, gamma, R2, type = "univariate", ...)

Arguments

n
Number of observations.

p
Number of variables.

q
An integer for univariate, a vector of 3 integers for bivariate and 3 or more for multivariate simulation (for details see Notes).

relpos
A list (vector in case of univariate simulation) of position of relevant component for predictor variables corresponding to each response.

gamma
A declining (decaying) factor of eigenvalues of predictors (X). Higher the value of gamma, the decrease of eigenvalues will be steeper.

R2
Vector of coefficient of determination (proportion of variation explained by predictor variable) for each relevant response components.

type
Type of simulation - univariate, bivariate and multivariate

...
Since this is a wrapper function to simulate univariate, bivariate or multivariate, it calls their respective function. This parameter should contain all the necessary arguements for respective simulations. See unisimrel, bisimrel and multisimrel

Value

A simrel object with all the input arguments along with following additional items.

X

Simulated predictors

Y

Simulated responses

W

Simulated predictor components

Z

Simulated response components

beta

True regression coefficients

beta0

True regression intercept

relpred

Position of relevant predictors

testX

Test Predictors

testY

Test Response

testW

Test predictor components

testZ

Test response components

minerror

Minimum model error

Xrotation

Rotation matrix of predictor (R)

Yrotation

Rotation matrix of response (Q)

type

Type of simrel object univariate or multivariate

lambda

Eigenvalues of predictors

SigmaWZ

Variance-Covariance matrix of components of response and predictors

SigmaWX

Covariance matrix of response components and predictors

SigmaYZ

Covariance matrix of response and predictor components

Sigma

Variance-Covariance matrix of response and predictors

RsqW

Coefficient of determination corresponding to response components

RsqY

Coefficient of determination corresponding to response variables

References

Sæbø, S., Almøy, T., & Helland, I. S. (2015). simrel—A versatile tool for linear model data simulation based on the concept of a relevant subspace and relevant predictors. Chemometrics and Intelligent Laboratory Systems, 146, 128-135.

Almøy, T. (1996). A simulation study on comparison of prediction methods when only a few components are relevant. Computational statistics & data analysis, 21(1), 87-107.