Extract various sigma matrices

Extract various sigma matrices

cov_mat(obj, which = c("xy", "zy", "zw"), use_population = TRUE)

Arguments

obj
A simrel object

which
A character string to specify which covariance matrix to extract, possible values are "xy", "zy" and "zw"

use_population
A boolean whether to use compute population values or to estimate from sample

Value

A matrix of covariances with column equals to the number of response and row equals to the number of predictors

Examples

set.seed(1983)
sobj <- multisimrel()
cov_mat(sobj, which = "xy", use_population = TRUE)

#>              [,1]        [,2]        [,3]        [,4]        [,5]
#>  [1,]  0.05473538  0.00000000  0.00000000  0.00000000  0.00000000
#>  [2,]  0.44858369  0.00000000  0.00000000  0.00000000  0.00000000
#>  [3,]  0.00000000 -0.10890699  0.00000000  0.00000000  0.10890699
#>  [4,]  0.00000000  0.02184183  0.00000000  0.00000000 -0.02184183
#>  [5,]  0.00000000  0.00000000  0.01324199 -0.01324199  0.00000000
#>  [6,]  0.00000000  0.19497385  0.00000000  0.00000000 -0.19497385
#>  [7,]  0.00000000  0.00000000  0.10910888 -0.10910888  0.00000000
#>  [8,]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [9,] -0.47857313  0.00000000  0.00000000  0.00000000  0.00000000
#> [10,]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [11,] -0.38466577  0.00000000  0.00000000  0.00000000  0.00000000
#> [12,]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [13,]  0.00000000 -0.10790869  0.00000000  0.00000000  0.10790869
#> [14,]  0.35991983  0.00000000  0.00000000  0.00000000  0.00000000
#> [15,]  0.00000000  0.00000000 -0.12235087  0.12235087  0.00000000

cov_mat(sobj, which = "xy", use_population = FALSE)

#>               [,1]         [,2]         [,3]         [,4]         [,5]
#>  [1,]  0.145263915  0.017631669 -0.032045772 -0.018564666  0.002962169
#>  [2,]  0.323151799 -0.057343006 -0.040733768 -0.036911362  0.012970942
#>  [3,] -0.009934655 -0.062877145  0.016093506  0.043792272  0.060131832
#>  [4,]  0.078225100 -0.001245780 -0.033014796 -0.035162014  0.029018231
#>  [5,]  0.019800746  0.004753679  0.019276631 -0.003389837  0.002513105
#>  [6,] -0.012727953  0.142092195 -0.006986039 -0.066301173 -0.172192504
#>  [7,] -0.013961217  0.019943101  0.095711969 -0.116314980  0.008524964
#>  [8,]  0.002648292 -0.001714722 -0.010480929 -0.007980862 -0.007653784
#>  [9,] -0.458636089  0.025047687  0.078075671  0.054159473 -0.016694564
#> [10,]  0.003724002  0.003838364  0.012438437 -0.003391220  0.006066926
#> [11,] -0.380194600  0.022665575  0.043690289  0.050744647 -0.020487909
#> [12,] -0.001929436 -0.004590749 -0.001657028  0.009683063 -0.004324354
#> [13,] -0.046783434 -0.071168913  0.014365823  0.046946504  0.088783345
#> [14,]  0.365710152 -0.012629579 -0.047273239 -0.051212973  0.021874649
#> [15,] -0.003294031 -0.021345598 -0.112795448  0.118332842 -0.008652385